refactor: encapsulate extended note controls
